refactor(form): split Form.svelte into a uniform section-component tree
Form.svelte (1330 lines) becomes a slim orchestrator owning only the
form-global concerns (field-value reconciliation, change-event debounce +
cancel-on-submit, autofocus, primary-action resolution, event emission).
Rendering moves to components/form/:

- One leaf component per section type; field leaves render their own
  label + control + error identically everywhere.
- A single Section.svelte dispatcher maps section type -> leaf, used by
  Form and by group items, so the routing table exists exactly once.
- GroupSection is a pure layout container: it renders its items through
  Form's recursive renderSection snippet (dependency inversion keeps the
  import graph acyclic; first-party circular deps fail the build).
- The group suppression contract (child labels not rendered, child
  errors below the row) and erroredFieldsOf are deleted: no production
  config ever used them, and children now render as ordinary sections.
  dialog-types.ts doc comment updated accordingly (types unchanged).

Tests split alongside: per-leaf test files drive components via props;
Form.test.ts keeps the orchestration suite (reconcile semantics, change
channel, autofocus, payload snapshots, group wiring).
