--- src/renderer/lib/components/Form.svelte
@@ -1,1329 +1,0 @@
-<!--
-  Form.svelte
-
-  Generic declarative form/section renderer.
-  Renders a DialogConfig's sections by type: text, progress, radio, dropdown,
-  table, input, group. Buttons live inside group sections: a button-only group
-  is a footer/action row, a group mixing a field with buttons attaches
-  side-flow buttons to that field's row. Every button click emits an action
-  event with the button's id and the full field-values snapshot.
-
-  Surface-agnostic: the wrapping surface (e.g. DialogView's modal card, or a
-  future panel shell) provides the chrome and text alignment. Form only lays the
-  sections out in a vertical flow.
-
-  Field sections (input/radio/dropdown) support an optional `label` (rendered
-  above the control) and an optional `error` (red helper text below the
-  control; the control is also marked invalid). The dropdown section is a
-  combobox (FilterableDropdown): a text input with a filtered suggestion list,
-  optionally accepting free text. `config.layout` switches the section layout:
-  "centered" (default) is the centered stack; "form" is left-aligned labeled
-  rows.
-
-  Dropdown sections support a `loading` flag: a spinner is overlaid at the
-  control's right edge while the backend fetches suggestions. The control
-  stays interactive and reports its value as usual.
--->
-<script lang="ts">
-  import type {
-    DialogConfig,
-    DialogButton,
-    DialogSection,
-    DialogUserEvent,
-    ProgressItem,
-  } from "@shared/dialog-types";
-  import { onMount, onDestroy, untrack } from "svelte";
-  import Icon from "./Icon.svelte";
-  import FilterableDropdown, {
-    type DropdownOption as FilterableOption,
-  } from "./FilterableDropdown.svelte";
-  import { sendDialogEvent } from "$lib/api";
-
-  type DropdownSection = Extract<DialogSection, { type: "dropdown" }>;
-
-  let formRef: HTMLElement | undefined;
-
-  interface Props {
-    dialogId: string;
-    config: DialogConfig;
-  }
-
-  const { dialogId, config }: Props = $props();
-
-  // Narrowed views of the shared section union (the individual section
-  // interfaces are intentionally not exported from dialog-types).
-  type GroupSection = Extract<DialogSection, { type: "group" }>;
-  type GroupItem = GroupSection["items"][number];
-  type ButtonItem = Extract<GroupItem, { type: "button" }>;
-  type FieldInput = Extract<DialogSection, { type: "input" }>;
-  type FieldLike = Extract<DialogSection | GroupItem, { type: "input" | "radio" | "dropdown" }>;
-
-  /**
-   * The element id a row label points at for a field: FilterableDropdown
-   * renders its text input as `${id}-input`; other controls use the id as-is.
-   */
-  function labelTargetOf(field: GroupItem): string {
-    return field.type === "dropdown" ? `${field.id}-input` : field.id;
-  }
-
-  /**
-   * All field sections of the config in declaration order — top-level
-   * input/radio/dropdown sections plus input/dropdown items nested in groups.
-   */
-  function fieldSectionsOf(cfg: DialogConfig): FieldLike[] {
-    const fields: FieldLike[] = [];
-    for (const section of cfg.sections) {
-      if (section.type === "input" || section.type === "radio" || section.type === "dropdown") {
-        fields.push(section);
-      } else if (section.type === "group") {
-        for (const item of section.items) {
-          if (item.type === "input" || item.type === "dropdown") {
-            fields.push(item);
-          }
-        }
-      }
-    }
-    return fields;
-  }
-
-  // Section layout: "centered" (default) keeps the centered stack; "form" lays
-  // fields out as left-aligned labeled rows with right-aligned actions.
-  const layout = $derived(config.layout ?? "centered");
-
-  // Track all field values (radio + dropdown + input) keyed by field id.
-  let fieldValues = $state<Record<string, string>>({});
-
-  // Dropdown display text keyed by field id: what the combobox input shows,
-  // which can differ from the reported value after a suggestion pick (the
-  // input displays the suggestion's label while the field reports its value).
-  let dropdownDisplay = $state<Record<string, string>>({});
-
-  /** A dropdown section's suggestions flattened across groups. */
-  function flatSuggestions(section: DropdownSection): readonly { value: string; label: string }[] {
-    return section.suggestions.flatMap((group) => group.items);
-  }
-
-  /** The label of the suggestion with this value, or the value itself. */
-  function suggestionLabel(section: DropdownSection, value: string): string {
-    return flatSuggestions(section).find((o) => o.value === value)?.label ?? value;
-  }
-
-  // Last backend-pushed `value` adopted per dropdown field. A deliberately
-  // non-reactive record: only the reconcile effect reads/writes it, to decide
-  // whether a config's `value` is new (adopt) or a re-send (preserve edits).
-  const adoptedValues: Record<string, string> = {};
-
-  // Reconcile field values whenever the config changes: rebuild the map so it
-  // mirrors the current field sections (dropping removed-field keys) while
-  // preserving values for fields that remain.
-  // - radio: keep the existing choice if still a valid option id, else the
-  //   first option's id.
-  // - dropdown (controlled): when the config carries a `value` the renderer
-  //   has not adopted yet, adopt it (strict mode falls back below when it
-  //   names no suggestion). Re-sends of the same value preserve user edits.
-  // - dropdown (freeText): like input — keep existing edits/picks, else seed
-  //   from initialValue on first sight.
-  // - dropdown (strict): keep the existing choice if still a valid suggestion
-  //   value; on first sight start at initialValue when it names a suggestion;
-  //   else the first suggestion's value. Display text follows the value
-  //   (suggestion label) unless the value is unchanged (preserving what the
-  //   user sees, e.g. typed free text).
-  // - input: keep existing edits/seeded value; otherwise seed from initialValue
-  //   on first sight (a later initialValue change does not re-seed).
-  // Existing values are read via untrack to avoid a write -> retrigger loop.
-  $effect(() => {
-    const next: Record<string, string> = {};
-    const nextDisplay: Record<string, string> = {};
-    for (const section of fieldSectionsOf(config)) {
-      if (section.type === "radio") {
-        const existing = untrack(() => fieldValues[section.id]);
-        const stillValid = existing !== undefined && section.options.some((o) => o.id === existing);
-        next[section.id] = stillValid ? existing : (section.options[0]?.id ?? "");
-      } else if (section.type === "dropdown") {
-        const existing = untrack(() => fieldValues[section.id]);
-        const pushed =
-          section.value !== undefined && section.value !== adoptedValues[section.id]
-            ? section.value
-            : undefined;
-        if (pushed !== undefined) {
-          adoptedValues[section.id] = pushed;
-        }
-        if (section.freeText) {
-          next[section.id] = pushed ?? existing ?? section.initialValue ?? "";
-        } else {
-          const isValid = (v: string | undefined): v is string =>
-            v !== undefined && flatSuggestions(section).some((o) => o.value === v);
-          next[section.id] = isValid(pushed)
-            ? pushed
-            : pushed === undefined && isValid(existing)
-              ? existing
-              : existing === undefined && pushed === undefined && isValid(section.initialValue)
-                ? section.initialValue
-                : (flatSuggestions(section)[0]?.value ?? "");
-        }
-        const existingDisplay = untrack(() => dropdownDisplay[section.id]);
-        nextDisplay[section.id] =
-          next[section.id] === existing && existingDisplay !== undefined && pushed === undefined
-            ? existingDisplay
-            : suggestionLabel(section, next[section.id]!);
-      } else if (section.type === "input") {
-        const existing = untrack(() => fieldValues[section.id]);
-        next[section.id] = existing ?? section.initialValue ?? "";
-      }
-    }
-    fieldValues = next;
-    dropdownDisplay = nextDisplay;
-  });
-
-  // ---- Field-change channel ----
-  // Per-field debounce timers, keyed by field id. A deliberately non-reactive
-  // plain record (never read in a reactive context). Emission is driven ONLY by
-  // user-interaction handlers (clicks, keystrokes), never by the reconcile
-  // effect above — so backend-driven config updates never re-emit (no loop).
-  const changeTimers: Record<string, ReturnType<typeof setTimeout>> = {};
-
-  /**
-   * Effective debounce (ms) for a field's change-event opt-in, or null when the
-   * field does not opt in. Default debounce per field type: continuous editing
-   * (input, dropdown typing) 200ms, radio (discrete) 0ms (immediate).
-   * Dropdown suggestion picks bypass this entirely (see handleDropdownSelect).
-   */
-  function changeDebounceMs(section: FieldLike): number | null {
-    const cfg = section.changeEvent;
-    if (!cfg) return null;
-    const fallback = section.type === "radio" ? 0 : 200;
-    if (cfg === true) return fallback;
-    return cfg.debounceMs ?? fallback;
-  }
-
-  /** Whether the dialog still has a field with this id. */
-  function hasField(fieldId: string): boolean {
-    return fieldSectionsOf(config).some((s) => s.id === fieldId);
-  }
-
-  /** Send a field-change event with the full keyed-values snapshot. */
-  function emitChange(fieldId: string): void {
-    // A config update may have removed the field while its timer was pending.
-    if (!hasField(fieldId)) return;
-    const event: DialogUserEvent = {
-      kind: "change",
-      dialogId,
-      fieldId,
-      data: getValues(),
-    };
-    sendDialogEvent(event);
-  }
-
-  /**
-   * Emit a change event for a field per its opt-in: immediately when the
-   * effective debounce is 0, otherwise on the trailing edge of a per-field
-   * timer. No-op for fields that did not opt in.
-   */
-  function scheduleChange(section: FieldLike): void {
-    const debounce = changeDebounceMs(section);
-    if (debounce === null) return;
-    const fieldId = section.id;
-    const pending = changeTimers[fieldId];
-    if (pending !== undefined) {
-      clearTimeout(pending);
-      delete changeTimers[fieldId];
-    }
-    if (debounce <= 0) {
-      emitChange(fieldId);
-      return;
-    }
-    changeTimers[fieldId] = setTimeout(() => {
-      delete changeTimers[fieldId];
-      emitChange(fieldId);
-    }, debounce);
-  }
-
-  /** Cancel all pending change timers (on submit or unmount). */
-  function cancelChangeTimers(): void {
-    for (const [fieldId, timer] of Object.entries(changeTimers)) {
-      clearTimeout(timer);
-      delete changeTimers[fieldId];
-    }
-  }
-
-  onDestroy(() => {
-    cancelChangeTimers();
-  });
-
-  /** Focus the control marked data-autofocus (after a tick for mounting). */
-  function focusAutofocusTarget(): void {
-    setTimeout(() => {
-      const target = formRef?.querySelector<HTMLElement>("[data-autofocus]");
-      target?.focus();
-    }, 0);
-  }
-
-  /** The id of the control carrying the autofocus flag, or null. */
-  function autofocusIdOf(cfg: DialogConfig): string | null {
-    for (const section of cfg.sections) {
-      if (section.type === "group") {
-        for (const item of section.items) {
-          if (item.autofocus) return item.id;
-        }
-      } else if (
-        (section.type === "input" || section.type === "dropdown" || section.type === "radio") &&
-        section.autofocus
-      ) {
-        return section.id;
-      }
-    }
-    return null;
-  }
-
-  // Auto-focus on mount: an explicit autofocus control wins, else the
-  // selected radio card (for keyboard navigation).
-  onMount(() => {
-    setTimeout(() => {
-      const explicit = formRef?.querySelector<HTMLElement>("[data-autofocus]");
-      if (explicit) {
-        explicit.focus();
-        return;
-      }
-      const selected = formRef?.querySelector("[aria-checked='true']") as HTMLElement | null;
-      selected?.focus();
-    }, 0);
-  });
-
-  // Focus follows when a config update MOVES the autofocus flag to a
-  // different control (e.g. picker button -> name field once a project
-  // exists). Re-sends of the same target never steal focus; the initial
-  // target is handled by onMount.
-  let lastAutofocusId: string | null | undefined;
-  $effect(() => {
-    const id = autofocusIdOf(config);
-    if (lastAutofocusId === undefined) {
-      lastAutofocusId = id;
-      return;
-    }
-    if (id !== null && id !== lastAutofocusId) {
-      focusAutofocusTarget();
-    }
-    lastAutofocusId = id;
-  });
-
-  /**
-   * Focus the textarea and either select the seeded text or place the caret
-   * at `cursorOffset`. Also re-focuses the textarea when Alt is released:
-   * Chromium's default Alt-up handler activates the window menu and pulls
-   * focus out of the WebContents, which after Alt+X+B would otherwise leave
-   * the dialog with no caret. (Electron #37336 prevents us from suppressing
-   * Alt at the keyDown layer.)
-   */
-  function seedCursor(
-    node: HTMLTextAreaElement,
-    params: {
-      initialValue: string | undefined;
-      cursorOffset: number | undefined;
-      selectInitialValue: boolean | undefined;
-    }
-  ): { destroy: () => void } | void {
-    if (params.initialValue === undefined) return;
-    const length = params.initialValue.length;
-    queueMicrotask(() => {
-      node.focus();
-      if (params.selectInitialValue) {
-        node.setSelectionRange(0, length);
-      } else if (params.cursorOffset !== undefined) {
-        const offset = Math.max(0, Math.min(params.cursorOffset, length));
-        node.setSelectionRange(offset, offset);
-      }
-      node.scrollTop = 0;
-    });
-
-    const onKeyUp = (e: KeyboardEvent): void => {
-      if (e.key !== "Alt") return;
-      // Re-focus only if focus was stolen (active element is body / null).
-      const active = document.activeElement;
-      if (active === node) return;
-      if (active && active !== document.body) return;
-      queueMicrotask(() => {
-        node.focus();
-      });
-    };
-    window.addEventListener("keyup", onKeyUp, true);
-    return {
-      destroy(): void {
-        window.removeEventListener("keyup", onKeyUp, true);
-      },
-    };
-  }
-
-  /**
-   * Enter in a single-line input activates the primary action (same as Enter
-   * in a dropdown). The action snapshot carries the typed value, so a pending
-   * debounced change is irrelevant. Attached as an action because the
-   * vscode-textfield custom element is opaque to the a11y template checks.
-   */
-  function submitOnEnter(node: HTMLElement): { destroy(): void } {
-    const handler = (e: KeyboardEvent): void => {
-      if (e.key === "Enter") {
-        e.preventDefault();
-        triggerPrimaryAction();
-      }
-    };
-    node.addEventListener("keydown", handler);
-    return {
-      destroy(): void {
-        node.removeEventListener("keydown", handler);
-      },
-    };
-  }
-
-  /** Snapshot every field's current value, keyed by field id. */
-  function getValues(): Record<string, string> {
-    const values: Record<string, string> = {};
-    for (const section of fieldSectionsOf(config)) {
-      values[section.id] = fieldValues[section.id] ?? "";
-    }
-    return values;
-  }
-
-  /**
-   * Map a dropdown section's suggestion groups onto FilterableDropdown's flat
-   * option list (a group's header becomes a non-selectable header entry).
-   */
-  function toFilterableOptions(section: DropdownSection): FilterableOption[] {
-    const result: FilterableOption[] = [];
-    section.suggestions.forEach((group, groupIndex) => {
-      if (group.header !== undefined && group.items.length > 0) {
-        result.push({ type: "header", label: group.header, value: `__header_${groupIndex}__` });
-      }
-      for (const item of group.items) {
-        result.push({ type: "option", label: item.label, value: item.value });
-      }
-    });
-    return result;
-  }
-
-  /**
-   * A suggestion was picked (or free text committed via Enter/Tab): report the
-   * picked value, display its label, and emit immediately — a pick is a
-   * discrete action, so it bypasses the typing debounce (cancelling any
-   * pending typing emit so the backend never sees stale text after the pick).
-   */
-  function handleDropdownSelect(section: DropdownSection, value: string): void {
-    fieldValues = { ...fieldValues, [section.id]: value };
-    dropdownDisplay = { ...dropdownDisplay, [section.id]: suggestionLabel(section, value) };
-    if (!section.changeEvent) return;
-    const pending = changeTimers[section.id];
-    if (pending !== undefined) {
-      clearTimeout(pending);
-      delete changeTimers[section.id];
-    }
-    emitChange(section.id);
-  }
-
-  /**
-   * The user typed in a dropdown. In free-text mode the typed text IS the
-   * field value (debounced change). In strict mode typing only filters the
-   * suggestion list — purely presentational, never reported.
-   */
-  function handleDropdownInput(section: DropdownSection, text: string): void {
-    if (!section.freeText) return;
-    fieldValues = { ...fieldValues, [section.id]: text };
-    dropdownDisplay = { ...dropdownDisplay, [section.id]: text };
-    scheduleChange(section);
-  }
-
-  /**
-   * Activate the form's primary button: the first button declared with an
-   * explicit variant "primary". Without one this does nothing — a positional
-   * fallback could land on a field-attached side-flow button.
-   */
-  function triggerPrimaryAction(): void {
-    const primaryButton = findPrimaryButton(config);
-    if (primaryButton) handleButton(primaryButton);
-  }
-
-  /**
-   * Activate the form's primary button as if clicked. Exposed for hosting
-   * surfaces that bind keyboard submit shortcuts (e.g. the panel shell's
-   * Cmd/Ctrl+Enter); same selection rule as Enter on a radio card. No-op
-   * without an explicit primary button.
-   */
-  export function submitPrimary(): void {
-    triggerPrimaryAction();
-  }
-
-  /** Handle a button click: emit an action event with the values snapshot. */
-  function handleButton(button: DialogButton): void {
-    if (button.disabled || button.busy) return;
-    // The action event carries the full snapshot, so any pending debounced
-    // change is redundant — cancel it to avoid a stray emit after submit.
-    cancelChangeTimers();
-    const event: DialogUserEvent = {
-      dialogId,
-      actionId: button.id,
-      data: getValues(),
-    };
-    sendDialogEvent(event);
-  }
-
-  /**
-   * The button Enter activates from a radio group: the first button declared
-   * with an explicit variant "primary". Without one, Enter does nothing — a
-   * positional fallback could land on a field-attached side-flow button.
-   */
-  function findPrimaryButton(cfg: DialogConfig): ButtonItem | undefined {
-    for (const section of cfg.sections) {
-      if (section.type !== "group") continue;
-      for (const item of section.items) {
-        if (item.type === "button" && item.variant === "primary") return item;
-      }
-    }
-    return undefined;
-  }
-
-  /** Resolved horizontal alignment of a group row (layout-natural default). */
-  function groupAlign(s: GroupSection): "left" | "center" | "right" {
-    return s.align ?? (layout === "form" ? "left" : "center");
-  }
-
-  /** A group's field items that carry an error (rendered below the row). */
-  function erroredFieldsOf(s: GroupSection): Exclude<GroupItem, ButtonItem>[] {
-    return s.items.filter(
-      (i): i is Exclude<GroupItem, ButtonItem> => i.type !== "button" && !!i.error
-    );
-  }
-
-  /** Parse text content for {badge:text} syntax. Returns segments. */
-  function parseTextContent(content: string): Array<{ type: "text" | "badge"; value: string }> {
-    const segments: Array<{ type: "text" | "badge"; value: string }> = [];
-    const regex = /\{badge:([^}]+)\}/g;
-    let lastIndex = 0;
-    let match: RegExpExecArray | null;
-
-    while ((match = regex.exec(content)) !== null) {
-      if (match.index > lastIndex) {
-        segments.push({ type: "text", value: content.slice(lastIndex, match.index) });
-      }
-      segments.push({ type: "badge", value: match[1]! });
-      lastIndex = regex.lastIndex;
-    }
-    if (lastIndex < content.length) {
-      segments.push({ type: "text", value: content.slice(lastIndex) });
-    }
-    return segments;
-  }
-
-  /** Get icon name for a progress item status. */
-  function getStatusIcon(status: ProgressItem["status"]): string {
-    switch (status) {
-      case "pending":
-        return "circle-outline";
-      case "running":
-        return "sync";
-      case "done":
-        return "check";
-      case "error":
-        return "error";
-    }
-  }
-
-  /** Get CSS class for a progress item status. */
-  function getStatusClass(status: ProgressItem["status"]): string {
-    switch (status) {
-      case "done":
-        return "status-done";
-      case "error":
-        return "status-error";
-      default:
-        return "";
-    }
-  }
-</script>
-
-<div class="form" class:layout-form={layout === "form"} bind:this={formRef}>
-  {#snippet fieldError(id: string, message: string)}
-    <vscode-form-helper id="{id}-error">
-      <span class="field-error">{message}</span>
-    </vscode-form-helper>
-  {/snippet}
-
-  {#snippet inputControl(s: FieldInput)}
-    {#if s.multiline}
-      <textarea
-        class="input-textarea"
-        class:errored={!!s.error}
-        class:fixed-rows={s.rows !== undefined}
-        id={s.id}
-        rows={s.rows}
-        placeholder={s.placeholder ?? ""}
-        disabled={s.disabled || undefined}
-        data-autofocus={s.autofocus || undefined}
-        aria-label={s.label ? undefined : (s.placeholder ?? "Text input")}
-        aria-invalid={s.error ? "true" : undefined}
-        aria-describedby={s.error ? `${s.id}-error` : undefined}
-        value={fieldValues[s.id] ?? ""}
-        use:seedCursor={{
-          initialValue: s.initialValue,
-          cursorOffset: s.cursorOffset,
-          selectInitialValue: s.selectInitialValue,
-        }}
-        oninput={(e) => {
-          fieldValues = { ...fieldValues, [s.id]: e.currentTarget.value };
-          scheduleChange(s);
-        }}
-      ></textarea>
-    {:else}
-      <vscode-textfield
-        class="input-textfield"
-        id={s.id}
-        invalid={s.error ? true : undefined}
-        placeholder={s.placeholder ?? ""}
-        disabled={s.disabled || undefined}
-        data-autofocus={s.autofocus || undefined}
-        aria-label={s.label ? undefined : (s.placeholder ?? "Text input")}
-        aria-invalid={s.error ? "true" : undefined}
-        aria-describedby={s.error ? `${s.id}-error` : undefined}
-        value={fieldValues[s.id] ?? ""}
-        oninput={(e: Event) => {
-          fieldValues = { ...fieldValues, [s.id]: (e.currentTarget as HTMLInputElement).value };
-          scheduleChange(s);
-        }}
-        use:submitOnEnter
-      ></vscode-textfield>
-    {/if}
-  {/snippet}
-
-  {#snippet dropdownControl(s: DropdownSection)}
-    <div class="dropdown-wrapper">
-      <FilterableDropdown
-        id={s.id}
-        options={toFilterableOptions(s)}
-        value={dropdownDisplay[s.id] ?? ""}
-        placeholder={s.placeholder ?? ""}
-        allowFreeText={s.freeText ?? false}
-        searchable={s.searchable ?? true}
-        disabled={s.disabled ?? false}
-        autofocus={s.autofocus ?? false}
-        debounceMs={0}
-        invalid={!!s.error}
-        describedBy={s.error ? `${s.id}-error` : undefined}
-        onSelect={(value) => handleDropdownSelect(s, value)}
-        onInput={(text) => handleDropdownInput(s, text)}
-        onEnter={() => triggerPrimaryAction()}
-      />
-      {#if s.loading}
-        <div class="dropdown-loading" role="status" aria-label="Loading options">
-          <Icon name="loading" spin />
-        </div>
-      {/if}
-    </div>
-  {/snippet}
-
-  {#snippet buttonControl(b: ButtonItem)}
-    <vscode-button
-      class:icon-button={!b.label}
-      secondary={b.variant === "secondary" || undefined}
-      disabled={b.disabled || b.busy || undefined}
-      data-autofocus={b.autofocus || undefined}
-      aria-label={b.label ?? b.title}
-      onclick={() => handleButton(b)}
-      {...b.title ? { title: b.title } : {}}
-    >
-      {#if b.icon}
-        <span class="button-icon" class:icon-only={!b.label}>
-          <Icon name={b.icon} spin={!!b.busy && !b.label} />
-        </span>
-      {/if}
-      {#if b.label}
-        {b.busy ? (b.busyLabel ?? b.label) : b.label}
-      {/if}
-    </vscode-button>
-  {/snippet}
-
-  {#each config.sections as section, sectionIndex (sectionIndex)}
-    {@const s = section}
-    {#if s.type === "text"}
-      {#if s.style === "heading"}
-        <h1 class="section-heading" class:has-icon={!!s.icon}>
-          {#if s.icon}
-            <span class="heading-icon" class:icon-error={s.icon === "error"}>
-              <Icon name={s.icon} size={20} />
-            </span>
-          {/if}
-          {#each parseTextContent(s.content) as seg, segIndex (segIndex)}
-            {#if seg.type === "badge"}<vscode-badge>{seg.value}</vscode-badge
-              >{:else}{seg.value}{/if}
-          {/each}
-        </h1>
-      {:else if s.style === "subtitle"}
-        <p class="section-subtitle">
-          {#each parseTextContent(s.content) as seg, segIndex (segIndex)}
-            {#if seg.type === "badge"}<vscode-badge>{seg.value}</vscode-badge
-              >{:else}{seg.value}{/if}
-          {/each}
-        </p>
-      {:else if s.style === "mono"}
-        <pre class="section-mono">{s.content}</pre>
-      {:else}
-        <p class="section-text">
-          {#each parseTextContent(s.content) as seg, segIndex (segIndex)}
-            {#if seg.type === "badge"}<vscode-badge>{seg.value}</vscode-badge
-              >{:else}{seg.value}{/if}
-          {/each}
-        </p>
-      {/if}
-    {:else if s.type === "progress"}
-      <div class="progress-container" role="status" aria-live="polite" aria-atomic="false">
-        {#each s.items as item, itemIndex (item.id)}
-          <div class="progress-row" class:progress-row-error={item.status === "error"}>
-            <div class="progress-row-header">
-              <span class="progress-status {getStatusClass(item.status)}">
-                <Icon name={getStatusIcon(item.status)} spin={item.status === "running"} />
-              </span>
-              <span class="progress-label">{item.label}</span>
-              {#if item.message && item.status !== "error"}
-                <span class="progress-message">{item.message}</span>
-              {/if}
-            </div>
-            {#if item.message && item.status === "error"}
-              <div class="progress-error-detail">{item.message}</div>
-            {/if}
-            {#if s.style !== "spinner"}
-              <div class="progress-bar-track">
-                {#if item.status === "running" && item.progress === undefined}
-                  <vscode-progress-bar indeterminate={true} aria-label="{item.label} progress"
-                  ></vscode-progress-bar>
-                {:else}
-                  {@const value =
-                    item.status === "done"
-                      ? 100
-                      : item.status === "pending"
-                        ? 0
-                        : (item.progress ?? 0)}
-                  <vscode-progress-bar
-                    {value}
-                    aria-label="{item.label} progress"
-                    aria-valuenow={value}
-                    aria-valuemin="0"
-                    aria-valuemax="100"
-                  ></vscode-progress-bar>
-                {/if}
-              </div>
-            {/if}
-          </div>
-          {#if itemIndex < s.items.length - 1}
-            <div class="progress-divider"></div>
-          {/if}
-        {/each}
-      </div>
-    {:else if s.type === "radio"}
-      <div class="form-field">
-        {#if s.label}
-          <vscode-label>{s.label}</vscode-label>
-        {/if}
-        <div
-          class="radio-cards"
-          class:errored={!!s.error}
-          role="radiogroup"
-          aria-label={s.label ?? "Selection"}
-          aria-describedby={s.error ? `${s.id}-error` : undefined}
-        >
-          {#each s.options as option, optionIndex (option.id)}
-            <button
-              type="button"
-              class="radio-card"
-              class:selected={fieldValues[s.id] === option.id}
-              role="radio"
-              aria-checked={fieldValues[s.id] === option.id}
-              tabindex={fieldValues[s.id] === option.id ? 0 : -1}
-              disabled={s.disabled || undefined}
-              data-option={option.id}
-              onclick={() => {
-                fieldValues = { ...fieldValues, [s.id]: option.id };
-                scheduleChange(s);
-              }}
-              onkeydown={(e) => {
-                const opts = s.options;
-                let targetIndex = -1;
-                if (e.key === "ArrowLeft" || e.key === "ArrowUp") {
-                  e.preventDefault();
-                  targetIndex = optionIndex === 0 ? opts.length - 1 : optionIndex - 1;
-                } else if (e.key === "ArrowRight" || e.key === "ArrowDown") {
-                  e.preventDefault();
-                  targetIndex = optionIndex === opts.length - 1 ? 0 : optionIndex + 1;
-                } else if (e.key === "Enter") {
-                  e.preventDefault();
-                  triggerPrimaryAction();
-                  return;
-                } else if (e.key === " ") {
-                  e.preventDefault();
-                  fieldValues = { ...fieldValues, [s.id]: option.id };
-                  scheduleChange(s);
-                  return;
-                }
-                if (targetIndex >= 0) {
-                  const targetId = opts[targetIndex]!.id;
-                  fieldValues = { ...fieldValues, [s.id]: targetId };
-                  scheduleChange(s);
-                  const container = e.currentTarget.closest(".radio-cards");
-                  setTimeout(() => {
-                    const card = container?.querySelector(
-                      `[data-option="${targetId}"]`
-                    ) as HTMLElement | null;
-                    card?.focus();
-                  }, 0);
-                }
-              }}
-            >
-              {#if option.icon}
-                <div class="radio-card-icon">
-                  <Icon name={option.icon} size={32} />
-                </div>
-              {/if}
-              <span class="radio-card-title">{option.label}</span>
-              <div class="radio-card-indicator">
-                {#if fieldValues[s.id] === option.id}
-                  <Icon name="circle-filled" size={16} />
-                {:else}
-                  <Icon name="circle-outline" size={16} />
-                {/if}
-              </div>
-            </button>
-          {/each}
-        </div>
-        {#if s.error}
-          <vscode-form-helper id="{s.id}-error">
-            <span class="field-error">{s.error}</span>
-          </vscode-form-helper>
-        {/if}
-      </div>
-    {:else if s.type === "dropdown"}
-      <div class="form-field">
-        {#if s.label}
-          <vscode-label for="{s.id}-input">{s.label}</vscode-label>
-        {/if}
-        {@render dropdownControl(s)}
-        {#if s.error}
-          {@render fieldError(s.id, s.error)}
-        {/if}
-      </div>
-    {:else if s.type === "input"}
-      <div class="form-field">
-        {#if s.label}
-          <vscode-label for={s.id}>{s.label}</vscode-label>
-        {/if}
-        {@render inputControl(s)}
-        {#if s.error}
-          {@render fieldError(s.id, s.error)}
-        {/if}
-      </div>
-    {:else if s.type === "group"}
-      {@const firstField = s.items.find((i) => i.type !== "button")}
-      {@const erroredFields = erroredFieldsOf(s)}
-      <div class="form-field">
-        {#if s.label}
-          <vscode-label for={firstField ? labelTargetOf(firstField) : undefined}
-            >{s.label}</vscode-label
-          >
-        {/if}
-        <div class="group-row align-{groupAlign(s)}" class:reverse={s.reverse}>
-          {#each s.items as item, itemIndex (itemIndex)}
-            {#if item.type === "input"}
-              <div class="group-field">{@render inputControl(item)}</div>
-            {:else if item.type === "dropdown"}
-              <div class="group-field">{@render dropdownControl(item)}</div>
-            {:else if item.type === "button"}
-              {@render buttonControl(item)}
-            {/if}
-          {/each}
-        </div>
-        {#each erroredFields as field (field.id)}
-          {@render fieldError(field.id, field.error ?? "")}
-        {/each}
-      </div>
-    {:else if s.type === "table"}
-      <div class="table-container">
-        {#if s.header}
-          <div class="table-header">
-            {#if s.headerIcon}
-              <span class="table-header-icon">
-                <Icon name={s.headerIcon} />
-              </span>
-            {/if}
-            <span>{s.header}</span>
-          </div>
-        {/if}
-        <div class="table-scroll">
-          <table>
-            <thead>
-              <tr>
-                {#each s.columns as col (col.key)}
-                  <th>{col.label}</th>
-                {/each}
-              </tr>
-            </thead>
-            <tbody>
-              {#each s.rows as row, rowIndex (rowIndex)}
-                <tr>
-                  {#each s.columns as col (col.key)}
-                    <td>{row[col.key] ?? ""}</td>
-                  {/each}
-                </tr>
-              {/each}
-            </tbody>
-          </table>
-        </div>
-      </div>
-    {/if}
-  {/each}
-</div>
-
-<style>
-  /* ---- Form root (inner layout) ---- */
-  /* Surface chrome + text alignment come from the wrapping surface (e.g. the
-     modal card in DialogView). Form only stacks sections vertically. */
-
-  .form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.75rem;
-    width: 100%;
-  }
-
-  /* Form-layout mode: left-aligned labeled rows. */
-  .form.layout-form {
-    align-items: stretch;
-    text-align: left;
-  }
-
-  /* Field wrapper: groups a field's label, control, and error tightly so the
-     error sits directly under its control (the 0.75rem .form gap only applies
-     between fields/sections). */
-  .form-field {
-    display: flex;
-    flex-direction: column;
-    gap: 0.25rem;
-    width: 100%;
-  }
-
-  /* Per-field validation error (slotted into <vscode-form-helper>). */
-  .field-error {
-    color: var(--ch-danger, #f14c4c);
-    font-size: 0.75rem;
-  }
-
-  /* ---- Text sections ---- */
-
-  .section-heading {
-    margin: 0;
-    font-size: 1.5rem;
-    font-weight: 500;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  .heading-icon {
-    display: flex;
-    align-items: center;
-  }
-
-  .icon-error {
-    color: var(--ch-danger);
-  }
-
-  .section-subtitle {
-    margin: 0;
-    font-size: 0.875rem;
-    opacity: 0.8;
-  }
-
-  .section-text {
-    margin: 0;
-    font-size: 0.875rem;
-  }
-
-  .section-mono {
-    margin: 0;
-    font-family: monospace;
-    font-size: 0.8rem;
-    white-space: pre-wrap;
-    word-break: break-word;
-    text-align: left;
-    width: 100%;
-    padding: 0.75rem;
-    background: color-mix(in srgb, var(--ch-foreground) 5%, transparent);
-    border-radius: var(--ch-radius-sm, 6px);
-  }
-
-  /* ---- Progress sections ---- */
-
-  .progress-container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    max-width: 400px;
-    padding: 0.5rem;
-    border-radius: var(--ch-radius-sm, 6px);
-  }
-
-  .progress-row {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    padding: 0.75rem;
-  }
-
-  .progress-row-error {
-    border-radius: var(--ch-radius-sm, 6px);
-  }
-
-  .progress-divider {
-    height: 1px;
-    margin: 0 0.75rem;
-    background: var(--ch-border);
-  }
-
-  .progress-row-header {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  .progress-status {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-shrink: 0;
-    width: 20px;
-    height: 20px;
-    opacity: 0.7;
-  }
-
-  .progress-status.status-done {
-    --vscode-icon-foreground: var(--ch-success, #89d185);
-    color: var(--ch-success, #89d185);
-    opacity: 1;
-  }
-
-  .progress-status.status-error {
-    --vscode-icon-foreground: var(--ch-danger, #f14c4c);
-    color: var(--ch-danger, #f14c4c);
-    opacity: 1;
-  }
-
-  .progress-label {
-    font-weight: 500;
-    flex-shrink: 0;
-  }
-
-  .progress-message {
-    margin-left: auto;
-    font-size: 0.875rem;
-    opacity: 0.7;
-  }
-
-  .progress-error-detail {
-    margin-left: 28px;
-    font-size: 0.75rem;
-    font-family: var(--vscode-editor-font-family, monospace);
-    color: var(--ch-danger, #f14c4c);
-    opacity: 0.85;
-    word-break: break-word;
-  }
-
-  .progress-bar-track {
-    width: 100%;
-    height: 4px;
-    background: var(--ch-input-background, rgba(255, 255, 255, 0.1));
-    border-radius: 2px;
-    overflow: hidden;
-  }
-
-  .progress-bar-track :global(vscode-progress-bar) {
-    width: 100%;
-    height: 100%;
-  }
-
-  .spinner-inline {
-    display: flex;
-    align-items: center;
-    height: 16px;
-  }
-
-  .spinner-inline :global(vscode-progress-ring) {
-    width: 16px;
-    height: 16px;
-  }
-
-  /* ---- Radio sections ---- */
-
-  .radio-cards {
-    display: flex;
-    /* The form-field wrapper is full-width; center the shrink-wrapped card
-       row in the default (centered) layout, like the pre-wrapper rendering. */
-    justify-content: center;
-    gap: 1rem;
-    margin: 0.5rem 0;
-  }
-
-  .form.layout-form .radio-cards {
-    justify-content: flex-start;
-  }
-
-  .radio-cards.errored {
-    outline: 1px solid var(--ch-danger, #f14c4c);
-    outline-offset: 6px;
-    border-radius: var(--ch-radius-md, 10px);
-  }
-
-  .radio-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 0.75rem;
-    width: 140px;
-    padding: 1.5rem 1rem;
-    border: 1px solid var(--ch-border);
-    border-radius: var(--ch-radius-md, 10px);
-    background: var(--ch-panel-background);
-    cursor: pointer;
-    transition:
-      border-color 0.15s ease,
-      background-color 0.15s ease;
-    color: inherit;
-    font-family: inherit;
-  }
-
-  .radio-card:hover {
-    border-color: var(--ch-focus-border);
-    background: var(--ch-accent-muted, var(--ch-list-hover-bg));
-  }
-
-  .radio-card:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    pointer-events: none;
-  }
-
-  .radio-card:focus {
-    outline: none;
-    border-color: var(--ch-focus-border);
-    box-shadow: 0 0 0 1px var(--ch-focus-border);
-  }
-
-  .radio-card.selected {
-    border-color: var(--ch-focus-border);
-    background: var(--ch-accent-muted, var(--ch-list-active-bg));
-  }
-
-  .radio-card-icon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 48px;
-    height: 48px;
-  }
-
-  .radio-card-title {
-    font-size: 1rem;
-    font-weight: 500;
-  }
-
-  .radio-card-indicator {
-    display: flex;
-    align-items: center;
-    opacity: 0.7;
-  }
-
-  .radio-card.selected .radio-card-indicator {
-    opacity: 1;
-    color: var(--ch-focus-border);
-  }
-
-  /* ---- Dropdown sections ---- */
-
-  .dropdown-wrapper {
-    position: relative;
-    width: 100%;
-  }
-
-  /* Loading spinner overlaid at the combobox input's right edge. The control
-     stays interactive (pointer-events: none). */
-  .dropdown-loading {
-    position: absolute;
-    right: 8px;
-    top: 50%;
-    transform: translateY(-50%);
-    display: flex;
-    align-items: center;
-    pointer-events: none;
-    color: var(--ch-foreground);
-    opacity: 0.7;
-  }
-
-  /* ---- Table sections ---- */
-
-  .table-container {
-    width: 100%;
-    margin-top: 0.5rem;
-    text-align: left;
-  }
-
-  .table-header {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    margin-bottom: 0.5rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-  }
-
-  .table-header-icon {
-    display: flex;
-    align-items: center;
-    color: var(--ch-warning);
-  }
-
-  .table-scroll {
-    max-height: 200px;
-    overflow-y: auto;
-    border: 1px solid var(--ch-border);
-    border-radius: var(--ch-radius-sm, 6px);
-  }
-
-  table {
-    width: 100%;
-    border-collapse: collapse;
-    font-size: 0.8rem;
-  }
-
-  th {
-    padding: 0.5rem 0.75rem;
-    text-align: left;
-    font-weight: 500;
-    background: color-mix(in srgb, var(--ch-foreground) 5%, transparent);
-    border-bottom: 1px solid var(--ch-border);
-    position: sticky;
-    top: 0;
-  }
-
-  td {
-    padding: 0.4rem 0.75rem;
-    border-bottom: 1px solid color-mix(in srgb, var(--ch-border) 50%, transparent);
-  }
-
-  tr:last-child td {
-    border-bottom: none;
-  }
-
-  /* ---- Input sections ---- */
-
-  .input-textarea {
-    width: 100%;
-    min-height: 30vh;
-    max-height: 60vh;
-    padding: 0.5rem;
-    font-family: inherit;
-    font-size: 0.875rem;
-    color: var(--ch-foreground);
-    /* Same surface as the other inputs (FilterableDropdown, vscode-textfield
-       both resolve to --vscode-settings-textInputBackground). */
-    background: var(--ch-input-bg);
-    border: 1px solid var(--ch-input-border);
-    border-radius: var(--ch-radius-sm, 6px);
-    resize: vertical;
-  }
-
-  .input-textarea:focus {
-    outline: none;
-    border-color: var(--ch-focus-border);
-  }
-
-  .input-textarea.errored {
-    border-color: var(--ch-danger, #f14c4c);
-  }
-
-  /* An explicit `rows` controls the initial height (user can still resize). */
-  .input-textarea.fixed-rows {
-    min-height: 0;
-  }
-
-  .input-textarea::placeholder {
-    color: var(--ch-foreground-dim, rgba(255, 255, 255, 0.5));
-  }
-
-  .input-textfield {
-    width: 100%;
-  }
-
-  /* ---- Group sections ---- */
-
-  /* Horizontal row of field controls and buttons. Field controls stretch
-     (via .group-field), buttons keep their natural size; `align` only shows
-     when no stretching field fills the row (e.g. a button-only footer). */
-  .group-row {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    width: 100%;
-  }
-
-  .group-row.align-left {
-    justify-content: flex-start;
-  }
-
-  .group-row.align-center {
-    justify-content: center;
-  }
-
-  .group-row.align-right {
-    justify-content: flex-end;
-  }
-
-  /* Visually reversed row (tab order keeps declaration order). The main axis
-     flips, so the justify mapping flips with it. */
-  .group-row.reverse {
-    flex-direction: row-reverse;
-  }
-
-  .group-row.reverse.align-right {
-    justify-content: flex-start;
-  }
-
-  .group-row.reverse.align-left {
-    justify-content: flex-end;
-  }
-
-  /* Rows mixing a field with buttons stretch the buttons to the field's
-     height (the old project-row look); button-only rows keep natural height. */
-  .group-row:has(.group-field) {
-    align-items: stretch;
-  }
-
-  .group-field {
-    flex: 1;
-    min-width: 0;
-  }
-
-  /* Icon rendered inside a labeled button gets breathing room before the
-     text; icon-only buttons don't need it. */
-  .button-icon {
-    display: inline-flex;
-    align-items: center;
-  }
-
-  .button-icon:not(.icon-only) {
-    margin-right: 0.3rem;
-  }
-
-  /* Icons slotted into buttons follow the button's text color instead of the
-     muted global --vscode-icon-foreground, which is barely visible on the
-     filled primary background. The button chrome itself is untouched. */
-  .group-row :global(vscode-button) {
-    --vscode-icon-foreground: var(--vscode-button-foreground, #ffffff);
-  }
-
-  .group-row :global(vscode-button[secondary]) {
-    --vscode-icon-foreground: var(--vscode-button-secondaryForeground, #cccccc);
-  }
-</style>
